% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{adaptive.MH}
\alias{adaptive.MH}
\title{adaptive.MH}
\usage{
adaptive.MH(sigma.Q, accept.rates, target = 0.1, update.size, tol = 0.8)
}
\arguments{
\item{sigma.Q}{proposal distribution variance parameter}

\item{accept.rates}{acceptance rate from previous iteration}

\item{target}{target acceptance rate}

\item{update.size}{size of update to be adjusted}

\item{tol}{tolerance level to determine if acceptance rate is too high}
}
\value{
nsamp number of samples with each row denoting binary vector
}
\description{
adaptive Metropolis Hastings to maintain target acceptance rate
}
