% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{adaptive_MH}
\alias{adaptive_MH}
\title{adaptive_MH}
\usage{
adaptive_MH(sigma_Q, accept_rates, target = 0.25, update_size, tol = 0.15)
}
\arguments{
\item{sigma_Q}{proposal distribution variance parameter for beta, delta, and eta}

\item{accept_rates}{acceptance rate from previous iteration}

\item{target}{target acceptance rate}

\item{update_size}{size of update to be adjusted}

\item{tol}{tolerance level to determine if acceptance rate is too high}
}
\value{
nsamp number of samples with each row denoting binary vector
}
\description{
adaptive Metropolis Hastings to maintain target acceptance rate
}
