% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_2.R
\name{add_2}
\alias{add_2}
\title{Add plus 2 to a number}
\usage{
add_2(x)
}
\arguments{
\item{x}{A number (integer)}
}
\value{
a number object
}
\description{
More information about the function but still short.
}
\examples{
add_2(2)
}
