% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{bayes.mfdur}
\alias{bayes.mfdur}
\title{bayes.mfdur}
\usage{
bayes.mfdur(formula, data = list(), N, burn, thin, w = c(1, 1, 1), m = 10,
  form = c("Weibull", "Exponential"), na.action = c("na.omit", "na.fail"))
}
\arguments{
\item{formula}{a formula in the form Y ~ X1 + X2... | C ~ Z1 + Z2 ... where Y is the duration until failure or censoring, and C is a binary indicator of observed failure.}

\item{data}{list object of data.}

\item{N}{number of MCMC iterations.}

\item{burn}{burn-in to be discarded.}

\item{thin}{thinning to prevent autocorrelation.}

\item{w}{size of the slice in the slice sampling for (betas, gammas, lambda). The default is c(1,1,1). This value may be changed by the user to meet one's needs.}

\item{m}{limit on steps in the slice sampling. The default is 10. This value may be changed by the user to meet one's needs.}

\item{form}{type of parametric model distribution to be used. Options are "Exponential" or "Weibull". The default is "Weibull".}

\item{na.action}{a function indicating what should happen when NAs are included in the data. Options are "na.omit" or "na.fail". The default is "na.omit".}
}
\description{
Markov Chain Monte Carlo (MCMC) to run Bayesian parametric MF model
}
