% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{BetaInEqC}
\alias{BetaInEqC}
\title{BetaInEqC}
\usage{
BetaInEqC(nIP, lambda.i, document)
}
\arguments{
\item{nIP}{total number of interaction patterns specified by the user}

\item{lambda.i}{stochastic intensity of specific sender "i"}

\item{document}{the id of specific edge of interest}
}
\value{
The vector of constants representing Beta part of each interaction pattern
}
\description{
Calculate Beta part of the equation used in multinomial sampling of C
}

