% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{GenerateDocs.PPC}
\alias{GenerateDocs.PPC}
\title{GenerateDocs.PPC}
\usage{
GenerateDocs.PPC(nDocs, node, vocabulary, nIP, K, alpha, mvec, betas, nvec, iJi,
  b, delta, currentC, netstat, base.edge, base.text, currentZ,
  word_type_topic_counts)
}
\arguments{
\item{nDocs}{number of documents to be generated}

\item{node}{nodelist containing the ID of nodes (ID starting from 1)}

\item{vocabulary}{all vocabularies used over the corpus}

\item{nIP}{total number of interaction patterns specified by the user}

\item{K}{total number of topics specified by the user}

\item{alpha}{Dirichlet concentration prior for document-topic distribution}

\item{mvec}{Dirichlet base prior for document-topic distribution}

\item{betas}{Dirichlet concentration prior for topic-word distribution}

\item{nvec}{Dirichlet base prior for topic-word distribution}

\item{iJi}{inferred latent receivers}

\item{b}{List of coefficients estimating the history effect for each IP}

\item{delta}{tuning parameter controlling the number of recipients}

\item{currentC}{topic-IP assignment}

\item{netstat}{which type of network statistics to use ("intercept", "dyadic", "triadic", "degree")}

\item{base.edge}{edges before 384 hours that is used to calculate initial history of interactions}

\item{base.text}{texts corresponding to base.edge}

\item{currentZ}{topic-token assignments}

\item{word_type_topic_counts}{word_type_topic_counts from inference}
}
\value{
generated edge and text, parameter b used to generate those, and base (if base == TRUE)
}
\description{
Generate a collection of documents according to the generative process of IPTM using Gibbs measure
}
