% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{GenerateDocs}
\alias{GenerateDocs}
\title{GenerateDocs}
\usage{
GenerateDocs(nDocs, node, vocab, nIP, K, n.d, alphas, beta, zeta, b, eta, delta,
  sigma_tau, cd_assignments = NULL, support, netstat, timestat,
  base.data = NULL, topic_token_assignments = NULL, backward = FALSE,
  base = FALSE, timeunit = 3600, tz = "America/New_York")
}
\arguments{
\item{nDocs}{number of documents to be generated}

\item{node}{vector of node id's (ID starting from 1)}

\item{vocab}{all vocabularies used over the corpus}

\item{nIP}{total number of interaction patterns}

\item{K}{total number of topics}

\item{n.d}{number of words in a document (fixed constant for now)}

\item{alphas}{Dirichlet concentration prior for document-topic distribution}

\item{beta}{Dirichlet concentration prior for topic-word distribution}

\item{zeta}{Dirichlet concentration prior for document-interaction pattern distribution}

\item{b}{coefficients for recipients}

\item{eta}{coefficients for timestamps}

\item{delta}{tuning parameter for the number of recipients}

\item{sigma_tau}{variance parameter for the timestamps}

\item{cd_assignments}{document-interaction pattern assignments}

\item{support}{support of latent recipients}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}

\item{timestat}{additional statistics to be used for timestamps other than netstat ("sender", "receiver", "timeofday", "dayofweek")}

\item{base.data}{edges before 384 hours that is used to calculate initial history of interactions}

\item{topic_token_assignments}{matrix of topic-token assignments}

\item{backward}{Logigal indicating whether we are generating backward samples (if FALSE -> forward)}

\item{base}{Logical indicating whether or not we are generating base edges (< 384)}

\item{timeunit}{hour (= 3600) or day (=3600*24) and so on}

\item{tz}{timezone such as EST and PST}
}
\value{
generated data including (author, recipients, timestamp, words)
}
\description{
Generate a collection of documents according to the generative process of IPTM
}
