% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{gibbs.measure.support}
\alias{gibbs.measure.support}
\title{gibbs.measure.support}
\usage{
gibbs.measure.support(n)
}
\arguments{
\item{n}{length of the vector to be sampled}
}
\value{
a 2^n x n binary matrix representing the support of the binary Gibbs measure in n elements
}
\description{
List out the support of Gibbs measure
}
