% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{GiR_stats}
\alias{GiR_stats}
\title{GiR_stats}
\usage{
GiR_stats(GiR_sample, K, currentC, vocabulary, forward = FALSE,
  backward = FALSE)
}
\arguments{
\item{GiR_sample}{one sample from generative process}

\item{K}{total number of topics specified by the user}

\item{currentC}{topic-IP assignment}

\item{vocabulary}{all vocabularies used over the corpus}

\item{forward}{Logigal indicating whether the sample is from forward sampling}

\item{backward}{Logigal indicating whether the sample is from backward sampling}
}
\value{
A vector of statistics calculated from one GiR sample
}
\description{
Calculate several statistics from samples generated from forward or backward sampling
}
