% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodbye.R
\name{goodbye}
\alias{goodbye}
\title{Print a Goodbye World message}
\usage{
goodbye()
}
\value{
None. This function is called for its side effects.
}
\description{
This function prints "Goodbye, world!" to the console.
}
\examples{
goodbye()
}
