% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{IPTM_predict.data}
\alias{IPTM_predict.data}
\title{IPTM_predict.data}
\usage{
IPTM_predict.data(D, O, R, edge, node, textlist, vocabulary, nIP, K, sigma_Q,
  alpha, mvec, betas, nvec, prior.b.mean, prior.b.var, prior.delta, out, n_B,
  n_d, burn, thinning, netstat, optimize = TRUE, initial)
}
\arguments{
\item{D}{Dth document the user wants to predict}

\item{O}{number of outer iterations of inference from which to generate predictions}

\item{R}{the number of iterations to sample predicted data within each outer iteration}

\item{edge}{list of document information with 3 elements (element 1 sender, element 2 receiver, element 3 time in unix.time format)}

\item{node}{nodelist containing the ID of nodes (ID starting from 1)}

\item{textlist}{list of text (length=number of documents in total) containing the words in each document}

\item{vocabulary}{all vocabularies used over the corpus}

\item{nIP}{total number of interaction patterns specified by the user}

\item{K}{total number of topics specified by the user}

\item{sigma_Q}{proposal distribution variance parameter for beta and delta}

\item{alpha}{Dirichlet concentration prior for document-topic distribution}

\item{mvec}{Dirichlet base prior for document-topic distribution}

\item{betas}{Dirichlet concentration prior for topic-word distribution}

\item{nvec}{Dirichlet base prior for topic-word distribution}

\item{prior.b.mean}{mean vector of b in multivariate normal distribution}

\item{prior.b.var}{covairance matrix of b in multivariate normal distribution}

\item{prior.delta}{parameter of delta in Normal prior}

\item{out}{size of outer iterations}

\item{n_B}{size of third inner iteration for updates of B}

\item{n_d}{size of third inner iteration for updates of delta}

\item{burn}{iterations to be discarded at the beginning of beta and delta chain}

\item{thinning}{the thinningning interval of beta and delta chain}

\item{netstat}{which type of network statistics to use ("intercept", dyadic", "triadic", "degree")}

\item{optimize}{to optimize alpha (Dirichlet concentration prior for document-topic distribution) or not (TRUE/FALSE)}

\item{initial}{list of initial values user wants to assign}
}
\value{
prediction output
}
\description{
posterior predictive experiments
}
