% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{lambda.post2}
\alias{lambda.post2}
\title{lambda.post2}
\usage{
lambda.post2(Y, eXB, alpha, C, lambda, a = 1, b = 1)
}
\arguments{
\item{Y}{response variable}

\item{eXB}{exponentiated vector of covariates times betas}

\item{alpha}{probability of true censoring}

\item{C}{censoring indicator}

\item{lambda}{current value of lambda}

\item{a}{shape parameter of gammas prior}

\item{b}{scale parameter of gammas prior}
}
\value{
log- posterior density of betas
}
\description{
log-posterior distribution of lambda
}
