% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg2-package.R
\docType{package}
\name{pkg2-package}
\alias{pkg2}
\alias{pkg2-package}
\title{pkg2: Example Package Two}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Author Name \email{example@example.com}

}
\keyword{internal}
