% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{PlotBetaIP}
\alias{PlotBetaIP}
\title{PlotBetaIP}
\usage{
PlotBetaIP(MCMCchain)
}
\arguments{
\item{MCMCchain}{a chain obtained using MCMC function}
}
\value{
Joint boxplot of the posterior summary of beta (should click for the locator)
}
\description{
Draw a boxplot of the MCMC chain of network statistics (beta) for each interaction pattern
}

