% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{hello}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{pkg1}{\code{\link[pkg1]{hello}}}
}}

