% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Schein}
\alias{Schein}
\title{Schein}
\usage{
Schein(Nsamp, nDocs, node, vocab, nIP, K, n.d, alphas, beta, zeta, prior.b,
  prior.delta, prior.eta, prior.tau, sigma.Q, Outer, Inner,
  netstat = c("dyadic"), timestat = c("timeofday", "dayofweek"), base.data,
  generate_PP_plots = TRUE)
}
\arguments{
\item{Nsamp}{number of GiR samples to be generated}

\item{nDocs}{number of documents to be generated per each sample}

\item{node}{vector of node id's (ID starting from 1)}

\item{vocab}{all vocabularies used over the corpus}

\item{nIP}{total number of interaction patterns}

\item{K}{total number of topics}

\item{n.d}{number of words in a document (fixed constant for now)}

\item{alphas}{Dirichlet concentration prior for document-topic distribution}

\item{beta}{Dirichlet concentration prior for topic-word distribution}

\item{zeta}{Dirichlet concentration prior for document-interaction patterndistribution}

\item{prior.b}{prior mean and covariance of b in multivariate Normal distribution}

\item{prior.delta}{prior mean and variance of delta in Normal distribution}

\item{prior.eta}{prior mean and covariance of eta in multivariate Normal distribution}

\item{prior.tau}{prior shape and scale parameter of sigma_tau in inverse-Gamma distribution}

\item{sigma.Q}{proposal distribution variance parameter}

\item{Outer}{size of outer iterations}

\item{Inner}{size of inner iteration for Metropolis-Hastings updates}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}

\item{timestat}{additional statistics to be used for timestamps other than netstat ("sender", "receiver", "timeofday", "dayofweek")}

\item{base.data}{artificial collection of documents to be used as initial state of history}

\item{generate_PP_plots}{Logical indicating whether to draw PP plots}
}
\value{
Forward and Backward samples and corresponding test results
}
\description{
Schein (preliminary GiR) test for the IPTM
}
