% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2013_doc.R
\docType{data}
\name{TA2013}
\alias{TA2013}
\title{Census results by Territorial Authority 2013}
\format{A data frame with 68 rows and 76 variables.}
\source{
\url{http://www3.stats.govt.nz/meshblock/2013/csv/2013_mb_dataset_Total_New_Zealand_CSV.zip}
}
\usage{
TA2013
}
\description{
Selected variables from the New Zealand 2013 census results by Territorial 
Authority (Districts and Cities)
}
\details{
\itemize{
  \item \code{TA}. Territorial Authority code
  \item \code{MeanBedrooms2013}. Mean number of bedrooms per dwelling. 
  \item \code{PropPrivateDwellings2013}. Proportion of dwellings that are private
  dwellings.  
  \item \code{PropSeparateHouse2013}. Proportion of occupied private dwellings
  that are a separate house (as opposed to appartments, two or more houses, etc)
  \item \code{NumberInHH2013} Mean usual number of residents in households
  for private dwellings.
  \item \code{PropMultiPersonHH2013} Proportion of occupied private dwellings
   that are 'other multi person household' ie not a single person and not one
   or more families
  \item \code{PropInternetHH2013} Proportion of occupied private dwellings
  that have access to the internet
  \item \code{PropNotOwnedHH2013} Proportion of occupied private dwellings 
  that are not owned or held in private trust for the occupiers
  \item \code{MedianRentHH2013} Median weekly rent for rented occupied private dwellings
  \item \code{PropLandlordPublic2013} Proportion of rented occupied private dwellings
  where the landlord is one of Housing New Zealand, Local Authority, City Council,
  or other state-owned corporation or enterprise
  \item \code{PropNoMotorVehicle2013} Proportion of occupied private dwellings 
  that have no motor vehicle for the household
  \item \code{PropAreChildren2013} Proportion of individuals aged under 15 years
  \item \code{PropSameResidence5YearsAgo2013} Proportion of individuals who
  lived in the same residence 5 years ago
  \item \code{PropOverseas5YearsAgo2013} Proportion of individuals who lived
  overseas 5 years ago
  \item \code{PropNZBorn2013} Proportion of individuals who were born in New
  Zealand
  \item \code{PropEuropean2013} Proportion of individuals with European ethnic group
  \item \code{PropMaori2013} Proportion of individuals with Maori ethnic group
  \item \code{PropPacific2013} Proportion of individuals with Pacific Peoples ethnic group
  \item \code{PropAsian2013} Proportion of individuals with Asia ethnic group
  \item \code{PropMale2013} Proportion of individuals who are male
  \item \code{PropFemale2013} Proportion of individuals who are female
  \item \code{Prop20to24_2013} Proportion of individuals aged 20 to 24  
  \item \code{Prop25to29_2013} Proportion of individuals aged 25 to 29  
  \item \code{Prop30to34_2013} Proportion of individuals aged 30 to 34  
  \item \code{Prop35to39_2013} Proportion of individuals aged 35 to 39  
  \item \code{Prop40to44_2013} Proportion of individuals aged 40 to 44  
  \item \code{Prop45to49_2013} Proportion of individuals aged 45 to 49  
  \item \code{Prop50to54_2013} Proportion of individuals aged 50 to 54  
  \item \code{Prop55to59_2013} Proportion of individuals aged 55 to 59  
  \item \code{Prop60to64_2013} Proportion of individuals aged 60 to 64  
  \item \code{Prop65AndOlder_2013} Proportion of individuals aged 65 or older  
  \item \code{PropNoReligion2013} Proportion of individuals who stated a religion
  with no religion   
  \item \code{PropSmoker2013} Proportion of individuals who stated smoking behaviour
  who smoked   
  \item \code{PropSeparated2013} Proportion of individuals who stated their status
  as separated
  \item \code{PropPartnered2013} Proportion of individuals who stated their status
  who had a partner
  \item \code{PropOwnResidence2013} Proportion of individuals who stated their status
  who were owner or part-owner of their residence
  \item \code{PropNoChildren2013} Proportion of individuals who stated their status
  who had no children
  \item \code{PropNoQualification2013} Proportion of individuals who stated their status
  with no educational qualification
  \item \code{PropBachelor2013} Proportion of individuals who stated their status
  whose highest qualification is a Bachelor degree
  \item \code{PropDoctorate2013} Proportion of individuals who stated their status
  whose highest qualification is a Doctorate
  \item \code{PropFTStudent2013} Proportion of individuals who stated their status
  who are in full time study
  \item \code{PropPTStudent2013} Proportion of individuals who stated their status
  who are in part time study
  \item \code{MedianIncome2013} Median individual income
  \item \code{PropSelfEmployed2013} Proportion of individuals who stated their status
  who received income from being self-employed or business
  \item \code{PropUnemploymentBenefit2013} Proportion of individuals who stated their status
  who received income from unemployment benefits
  \item \code{PropStudentAllowance2013} Proportion of individuals who stated their status
  who received income from student allowance
  \item \code{PropFullTimeEmployed2013} Proportion of individuals resident who are
  employed full time
  \item \code{PropPartTimeEmployed2013} Proportion of individuals resident who are
  employed part time
  \item \code{PropUnemployed2013} Proportion of individuals resident who are
  unemployed
  \item \code{PropEmployee2013} Proportion of individuals resident who are
  paid employees
  \item \code{PropEmployer2013} Proportion of individuals resident who are
  employers
  \item \code{PropSelfEmployedNoEmployees} Proportion of residents resident
  who are self employed with no employees
  \item \code{PropManagers2013} Proportion of residents who stated an
  occupation who are Managers
  \item \code{PropProfessionals2013} Proportion of residents who stated an 
  occupation who are Professionals
  \item \code{PropTrades2013} Proportion of residents who stated an occupation
  who are Technicians and Trades Workers
  \item \code{PropLabourers2013} Proportion of residents who stated an 
  occupation who are Labourers
  \item \code{PropAgForFish2013} Proportion of residents who stated an
  industry who work in Agriculture, Forestry and Fishing
  \item \code{PropPubAdmin2013} Proportion of residents who stated an 
  industry who work in Public Administration and Safety
  \item \code{PropFinServices2013} Proportion of residents who stated an
  industry who work in Financial and Insurance Services
  \item \code{PropProfServices2013} Proportion of residents who stated an
  industry who work in Professional, Scientific and Technical Services
  \item \code{PropWorked1_9hours2013} Proportion of residents aged 15+ who 
  stated working hours who worked from 1 to 9 hours per week
  \item \code{PropWorked10_19hours2013} Proportion of residents aged 15+ who 
  stated working hours who worked from 10 to 19 hours per week
  \item \code{PropWorked20_29hours2013} Proportion of residents aged 15+ who 
  stated working hours who worked from 20 to 29 hours per week
  \item \code{PropWorked30_39hours2013} Proportion of residents aged 15+ who 
  stated working hours who worked from 30 to 39 hours per week
  \item \code{PropWorked40_49hours2013} Proportion of residents aged 15+ who 
  stated working hours who worked from 40 to 49 hours per week
  \item \code{PropWorked50_59hours2013} Proportion of residents aged 15+ who 
  stated working hours who worked from 50 to 59 hours per week
  \item \code{PropWorkedOver60hours2013} Proportion of residents aged 15+ who 
  stated working hours who worked over 60 hours per week
  \item \code{PropWorkedHome2013} Proportion of residents who worked at home
  \item \code{PropPublicTransport2013} Proportion of residents who take
  the bus or train to work
  \item \code{PropWalkJogBike2013} Proportion of residents who walked, jogged
  or rode a bicyle to work
  \item \code{PropNoUnpaidActivities2013} Proportion of residents who perform
  no unpaid activities
  \item \code{TA2013_NAM} Territorial Authority name
  \item \code{NZTM2000Easting} Easting NZTM2000 coordinate of the centroid
  \item \code{NZTM2000Northing} Northing NZTM2000 coordinate of the centroid
  \item \code{WGS84Longitude} Longitude of the centroid   
  \item \code{WGS84Latitude} Latitude of the centroid
}
Variables have been selected in a somewhat ad hoc fashion.  For example, not
all forms of travelling to work are shown; the same applies to occupations,
industries, etc.  Generally the focus has been on place of residence than
of work.  The aim is to provide a convenient collection of variables rather 
than to be comprehensive.

See the source (below) for the definitive and complete form of the original data 
and its metadata. The exact variables used in the creation of this dataframe 
can be seen at \url{https://github.com/ellisp/nzelect/blob/master/prep/import_census.R}
The census data was joined to data from the relevant shapefile by a method seen at
\url{https://github.com/ellisp/nzelect/blob/master/prep/add_locations_census.R}
}
\examples{
head(TA2013)
}
\seealso{
\code{\link{REGC2013}}
}
\keyword{datasets}

