% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{TopicInEqC}
\alias{TopicInEqC}
\title{TopicInEqC}
\usage{
TopicInEqC(nIP, K, currentZ, alpha, mvec, document)
}
\arguments{
\item{nIP}{total number of interaction patterns specified by the user}

\item{K}{total number of topics specified by the user}

\item{currentZ}{current state of the assignment of topics}

\item{alpha}{Dirichlet concentration prior for topic distribution}

\item{mvec}{Dirichlet base prior for topic distribution}

\item{document}{the id of specific edge of interest}
}
\value{
The vector of constants representing topic part of each interaction pattern
}
\description{
Calculate topic part of the equation used in multinomial sampling of C
}

