% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{PlotTopic}
\alias{PlotTopic}
\title{PlotTopic}
\usage{
PlotTopic(Zchain, K)
}
\arguments{
\item{Zchain}{summary of Z obtained using MCMC function}

\item{K}{total number of topics specified by the user}
}
\value{
Barplot of the topic distribution
}
\description{
Draw a barplot of the topic distributions without considering interaction patterns
}
